Clean up k6 folder and add to CI (#5419)
