+++ k6/.lintstagedrc.js
@@ -1,0 +1,6 @@
+const baseConfig = require('../.lintstagedrc.js');
+
+module.exports = {
+  ...baseConfig,
+  '*.js': ['eslint --fix'],
+};
