+++ k6/.node-version
@@ -1,0 +1,1 @@
+v18
