+++ k6/.prettierrc.yml
@@ -1,0 +1,7 @@
+arrowParens: always
+semi: true
+singleAttributePerLine: true
+singleQuote: true
+trailingComma: all
+plugins:
+  - prettier-plugin-organize-imports
