+++ k6/eslint.config.mjs
@@ -1,0 +1,11 @@
+import pluginJs from '@eslint/js';
+import eslintConfigPrettier from 'eslint-config-prettier';
+import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended';
+import globals from 'globals';
+
+export default [
+  { languageOptions: { globals: { ...globals.node } } },
+  pluginJs.configs.recommended,
+  eslintConfigPrettier,
+  eslintPluginPrettierRecommended,
+];
