+++ k6/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "121-service-k6-tests",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "directories": {
+    "test": "tests"
+  },
+  "scripts": {
+    "lint": "eslint .",
+    "fix": "npm run lint -- --fix",
+    "test": "dotenv -e ../services/.env -- k6 run"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@eslint/js": "^9.4.0",
+    "eslint": "^9.4.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-promise": "^6.2.0",
+    "globals": "^15.4.0",
+    "prettier-plugin-organize-imports": "^3.2.4"
+  },
+  "dependencies": {
+    "dotenv-cli": "^7.4.2"
+  }
+}
