+++ package.json
@@ -34,17 +34,9 @@
     "prepare": "husky"
   },
   "devDependencies": {
-    "@types/k6": "^0.51.0",
     "eslint-plugin-no-relative-import-paths": "^1.5.4",
     "husky": "^9.0.11",
     "lint-staged": "^15.2.2",
-    "prettier": "^3.2.5",
-    "ts-loader": "^9.5.1",
-    "typescript": "^5.4.5",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
-  },
-  "dependencies": {
-    "dotenv": "^16.4.5"
+    "prettier": "^3.2.5"
   }
 }
