+++ services/.env.example
@@ -285,6 +285,3 @@ MOCK_COMMERCIAL_BANK_ETHIOPIA=TRUE
 BASE_URL=http://localhost:8888
 AZURE_DEV_URL=https://dev.azure.com/redcrossnl
 AZURE_DEVOPS_TOKEN=
-
-# K6 Load Testing configuration
-API_BASE_URL=http://localhost:3000
